Move gelu matcher to its own pass and add test
