+++ src/simplify_algebra.cpp
@@ -851,63 +851,6 @@ struct find_div_const
     }
 };
 
-struct find_gelu_erf
-{
-    static auto match_div()
-    {
-        return match::name("div")(
-            args(match::any().bind("x"), match::skip_broadcasts(match::is_constant())));
-    }
-
-    static auto match_mul1()
-    {
-        return match::name("mul")(
-            args(match::any().bind("x"), match::skip_broadcasts(match::name("recip"))));
-    }
-
-    static auto match_erf() { return match::name("erf")(match::arg(0)(match_div())); }
-
-    static auto match_add2()
-    {
-        return match::name("add")(
-            args(match_erf(), match::skip_broadcasts(match::has_value(1.0f))));
-    }
-
-    static auto match_add1()
-    {
-        return match::name("add")(
-            args(match::skip_broadcasts(match::is_constant()), match::name("dot")));
-    }
-
-    static auto match_mul2() { return match::name("mul")(match::args(match_add1(), match_add2())); }
-
-    auto matcher() const
-    {
-        return match::name("mul")(
-            match::args(match_mul2(), match::skip_broadcasts(match::has_value(0.5f))));
-    }
-
-    void apply(module& m, const match::matcher_result& r) const
-    {
-        auto ins = r.result;
-        auto x   = r.instructions["x"];
-
-        auto lit = m.add_literal(literal{shape{x->get_shape().type()}, {1.702f}});
-        auto mul = m.insert_instruction(
-            ins, make_op("multibroadcast", {{"out_lens", x->get_shape().lens()}}), lit);
-        mul      = m.insert_instruction(ins, make_op("mul"), x, mul);
-        auto sig = m.insert_instruction(ins, make_op("neg"), mul);
-        auto one = m.add_literal(literal{shape{x->get_shape().type()}, {1.0f}});
-        one      = m.insert_instruction(
-            ins, make_op("multibroadcast", {{"out_lens", x->get_shape().lens()}}), one);
-        sig = m.insert_instruction(ins, make_op("exp"), sig);
-        sig = m.insert_instruction(ins, make_op("add"), one, sig);
-        sig = m.insert_instruction(ins, make_op("div"), one, sig);
-        sig = m.insert_instruction(ins, make_op("mul"), x, sig);
-        m.replace_instruction(ins, sig);
-    }
-};
-
 struct find_sub_const
 {
     auto matcher() const
@@ -1106,7 +1049,6 @@ void simplify_algebra::apply(module& m) const
                             find_mul_slice_conv{},
                             find_mul_add{},
                             find_div_const{},
-                            find_gelu_erf{},
                             find_sub_const{},
                             find_rsqrt{},
                             find_concat_op{},
